Rename: 모델 파일 구분을 위해 이름에 "model" 추가
